requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
assemblyai==0.48.4
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
distro==1.9.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.47.0
google-genai==1.60.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.1
narwhals==2.15.0
numpy==2.4.1
packaging==26.0
pandas==2.3.3
pillow==12.1.0
postgrest==2.27.2
propcache==0.4.1
protobuf==6.33.4
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
realtime==2.27.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
storage3==2.27.2
streamlit==1.53.1
StrEnum==0.4.15
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
yarl==1.22.0
yt-dlp==2025.12.8
